refactor(pipeline): extract shared email fixtures (asIso/emailToMessage) to testing/email-fixtures
Dedupes the byte-identical helpers across email-pipeline.test.ts and parquet-email.test.ts.
Test-only, not re-exported from testing/index.ts, so @dxos/types stays dev-only.
