+++ packages/core/compute/pipeline/src/testing/parquet-email.test.ts
@@ -9,10 +9,11 @@ import { join } from 'node:path';
 import { describe, test } from 'vitest';
 
 import { EffectEx } from '@dxos/effect';
-import { type ContentBlock, Message } from '@dxos/types';
+import { Message } from '@dxos/types';
 
 import * as Pipeline from '../Pipeline';
 import * as Stage from '../Stage';
+import { emailToMessage } from './email-fixtures';
 import { captureSink, scriptedSource } from './index';
 import { type ParquetRow, parquetSource } from './parquet';
 
@@ -104,27 +105,6 @@ describe('email parquet → Message', () => {
     });
   });
 });
-
-const asIso = (value: unknown): string => (value instanceof Date ? value : new Date(String(value))).toISOString();
-
-// Map one email row (see the dataset's `dataset_info` schema) to a Message carrying the body as a
-// text block. Test-only: keeps the generic @dxos/pipeline package free of an @dxos/types dependency.
-const emailToMessage = (row: ParquetRow): Message.Message => {
-  const block: ContentBlock.Text = { _tag: 'text', text: String(row.body ?? '') };
-  return Message.make({
-    created: asIso(row.date),
-    sender: { email: String(row.from ?? '') },
-    blocks: [block],
-    properties: {
-      messageId: row.message_id,
-      subject: row.subject,
-      to: row.to,
-      cc: row.cc,
-      bcc: row.bcc,
-      fileName: row.file_name,
-    },
-  });
-};
 
 // A pipeline stage that maintains a running count of emails per sender, emitting the updated count
 // for each email's sender as it flows through (a stateful scan; the counter map is bounded by the
